# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API YocoShort",
        "description": "Documentación oficial de YocoShort",
        "contact": {
            "email": "sisabuestandavidesteban@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.yocoshort.com",
            "description": "Servidor de Producción en Render"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local para desarrollo"
        }
    ],
    "paths": {
...
        "/api/domain": {
            "get": {
                "tags": [
                    "Domains"
                ],
                "summary": "Obtener configuración de mi dominio",
                "operationId": "d7f46227e6e2d3fe0963a7f457f44aa5",
                "responses": {
                    "200": {
                        "description": "Información del dominio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "subdomain": {
                                            "type": "string",
                                            "example": "mi-marca"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No tienes dominio configurado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Domains"
                ],
                "summary": "Crear o actualizar mi subdominio",
                "operationId": "af09a62f07cee038723fe722f30b0c7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "subdomain": {
                                        "description": "Min 3, Max 20 caracteres",
                                        "type": "string",
                                        "example": "mi-marca"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dominio configurado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Subdominio configurado exitosamente."
                                        },
                                        "domain": {
                                            "type": "string",
                                            "example": "mi-marca.yocoshort.com"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Subdominio reservado, duplicado o inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Domains"
                ],
                "summary": "Eliminar mi subdominio",
                "operationId": "d6e6eead7bfdf6d0f067fabfb8743879",
                "responses": {
                    "200": {
                        "description": "Dominio eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dominio eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe dominio para eliminar"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/google/callback": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Callback de Google",
                "description": "Maneja la respuesta de Google. Si el login es exitoso, redirige al Frontend con el token en la URL.",
                "operationId": "e9b82bcbf1bf4ce865aa533070f20eb8",
                "responses": {
                    "302": {
                        "description": "Redirección al Dashboard del Frontend (?token=...)"
                    },
                    "500": {
                        "description": "Error en el servidor o cancelación del usuario"
                    }
                }
            }
        },
        "/api/long-url": {
            "post": {
                "tags": [
                    "Public Shortener"
                ],
                "summary": "Acortar enlace publico sin loggeo.",
                "description": "Crea un enlace corto temporal sin necesidad de registro.",
                "operationId": "6c905e92fb3f6464f0023a9d3b3b50a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "long_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://youtube.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enlace creado o renovado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "short_url": {
                                            "type": "string",
                                            "example": "http://short.gy/AbCd12"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "URL inválida o muy larga"
                    }
                }
            }
        },
...
        "/api/cleanup": {
            "delete": {
                "tags": [
                    "System"
                ],
                "summary": "Limpiar enlaces expirados (Cronjob)",
                "operationId": "2ab893d078709edf8cb61eeda3104622",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Clave secreta del Cronjob",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Limpieza exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "alive"
                                        },
                                        "deleted_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (Clave incorrecta)"
                    }
                }
            }
        },
...
    },
...
}